% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{simplify_topology}
\alias{simplify_topology}
\title{Remove all 2-degree nodes from a graph}
\usage{
simplify_topology(graph, protected_nodes = NULL)
}
\arguments{
\item{graph}{An undirected \code{\link[igraph:igraph]{igraph}} object}

\item{protected_nodes}{Indices of nodes that should NOT be removed}
}
\value{
An undirected \code{\link[igraph:igraph]{igraph}} object with no 2-degree
nodes
}
\description{
New edges will inherit the attributes of whatever edge originated from the
head node that remains.
}
